Stop minifying review app CSS
cssnano-preset-default includes a number of transformations that meaningfully change the CSS, including [1]:

- sorting declarations (css-declaration-sorter)
- transforming colours (postcss-colormin)
- converting values (postcss-convert-values)
- merging longhand properties to shorthand (postcss-merge-longhand)
- merging rules (postcss-merge-rules)
- transforming fonts (postcss-minify-font-values)

We are generally very intentional with the way that we author our CSS. We want the review app to reflect the code we’re writing as closely as possible.

Allowing postcss to meaningfully change the CSS is confusing, and in some cases may mislead or mask problems.

As a more tangible example, this configuration is currently masking the fact that `govuk-tint` and `govuk-shade` output colours defined using `rgb()` with floats which are not supported in older browsers [2] [3]. The `postcss-colormin` plugin is converting them back to hex values in the review app.

Remove cssnano from the review app’s postcss config and uninstall.

This doesn’t affect the minified CSS shipped in the govuk-frontend package, which is processed using the separate postcss config in the package folder.

[1]: https://www.npmjs.com/package/cssnano-preset-default#plugins
[2]: https://caniuse.com/mdn-css_types_color_rgb_float_values
[3]: https://github.com/alphagov/govuk-frontend/issues/5988