Force skip link to be focused for screenshot
Make the ‘with focus’ example of the skip link visible in the review app and use it for screenshots.

This example uses the `:focus` ‘companion class’ we get from postcss-pseudo-classes.

Unfortunately the `govuk-visually-hidden-focusable` mixin uses a `&:not(:active):not(:focus)` which isn’t correctly handled by the postcss-pseudo-classes plugin, so we need to modify the rule slightly to prevent it from hiding the skip link.