Allow examples to set page template options
We currently allow each example to set an array of `previewLayoutModifiers` which affect how the example is rendered in the review app (and thus in our visual regression tests).

This is limited because:

- it assumes that the modifier will exist as a modifier class of `app-template__body`
- it can only be used to add classes to the body

Replace with a generic way to set ‘context’ for the example, which is then passed to `env.render`.

This will allow us to set e.g. `htmlClasses` on the page template using the same approach, allowing us to add examples that use the rebrand.