Remove getFragmentForUrl helper
Everywhere we’re currently using this helper we’re taking the URL from an `HTMLAnchorElement`.

Rather than pass the URL to `getFragmentForUrl` as a string and splitting on `#`, we can get the hash of the target URL using `HTMLAnchorElement.prototype.hash`.

The only thing that then needs to be done is to remove the leading `#`, which we can do with a simple replace that I don’t think is worth abstracting.