Add failing test for file upload with empty label
When using the JavaScript ‘enhanced’ version of the file upload component, we currently have a failure state where the native input is hidden _before_ we throw an error if no label is present.

We should throw defer hiding the native input until after we’re done with these checks.

Add a failing test that covers the issue as it’s described in #6104.