Fix failure state for file upload with empty label
When using the JavaScript ‘enhanced’ version of the file upload component, we currently have a failure state where the native input is hidden _before_ we throw an error if no label is present, resulting in neither the input nor the enhanced version being rendered.

Defer hiding the native input until after we’re done with these checks.