Move bundler-integrations workspace
Dependabot seems to have trouble with the bundler-integrations workspace and our hypothesis is that's got something to do with the fact that it sits in the github workflows directory.

Move it out the shared directory (as a best option - even though it's not _really_ shared) and see how that affects things.