Use optional catch binding for accordion
We ‘don’t care’ if setting or reading from session storage in the Accordion throws an error (which it can do if e.g. the user has configured the browsers to prevent the page from persisting data [1]).

As such we don’t do anything with the error passed into the catch block, which means we have an unused variable currently being flagged by the @typescript-eslint/no-unused-vars eslint rule.

I’ve checked that Babel transpiles optional catch binding based on our target browserslist, so we can safely disable the es-x/no-optional-catch-binding rule and use optional catch binding for the try / catch blocks in the Accordion’s javascript.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Window/sessionStorage#exceptions