Fix hover state of pagination links in Safari
The previous and next links within the pagination component include a nested span with the ‘Previous’ or ‘Next’ text in it, alongside the svg for the appropriate arrow.

In Safari, the span is not inheriting the `text-decoration-thickness` property from the parent link (correctly, we believe [^1]), so we need to explicitly set `text-decoration-thickness: inherit;` on the span so that it displays correctly.

Fixes #5335.

[^1]: https://github.com/alphagov/govuk-frontend/issues/5335#issuecomment-2361533572