Run percy on merges to relevant branches.
The tests workflow triggers on pull_request, push (to main, support and feature branches), and workflow dispatch.

If we've manually dispatched the Percy workflow, we already run Percy by default.

Now we also pass a trigger_event to the workflow and run Percy on a push - which will always be to one of the relevant branches.

This felt cleaner than checking the branches within the workflow's script.