[SPIKE] Lint usage of govuk-media-query mixin
Exploration into adding a custom Stylelint rule to check that we provide
a breakpoint to the `govuk-media-query` mixin.

Adds a new `@govuk-frontend/stylelint-plugin` workspaces in the repository,
hosting a Stylelint plugins with our custom rules.

The package is then referenced in our `stylelint.config.js` file
for use on the files inside `govuk-frontend` (as a test).

Code is very spike-y and untested. It was just an experiment to explore
the cost of having a little bit of linting checking our current usage of
`govuk-media-query`. It could be seen as a kind of test suite for #6246.

It does show that should we want it, we could have our own linting rules,
which could be helpful either for testing more opinionated choices about
the Sass or CSS features we use.

We could also explore making public linting rules on how best to use GOV.UK Frontend.