Fix failing accessible name tests
Some of our tests for accessible names have started failing after a Puppeteer update. We aren't totally sure why this is, but it appears that the nodes we're testing are now considered 'uninteresting' and are being pruned from the node list by Puppeteer.

This commit sets `interestingOnly: false` so that Puppeteer doesn't prune the node list prior to running the test.

https://pptr.dev/api/puppeteer.accessibility.snapshot#remarks