Expect errors from failing moduleResolution
We use Node for our typescript moduleResolution. This defaults to Node10, which means it doesn't support subpath exports.

html-validate happens to use subpath exports in v10.0.0, and this causes TypeScript to complain.

At some point, we probably should update our moduleResolution to nodenext , but doing that right now creates a lot of errors, so we're kicking the can down the road here and solving this by expecting the moduleResolution errors.

If this actually breaks our implementation, the tests themselves should fail, so the risk is small.