Bind express to localhost rather than all IPs
By default, `server.listen` will accept connections on 0.0.0.0 [1], which means you can access the review app using its local IP address from the same LAN (and, if there is a public IP, from the internet).

We have cases where we may want to allow users to connect to the review app externally (e.g. testing on a mobile device whilst connected to the same LAN as a developer machine) but we can use the BrowserSync URLs for this. There is no need to allow external connections from the express app itself.

Configure the Express app to bind to localhost instead, which only accepts connections from the same computer.

[1]: https://nodejs.org/api/net.html#serverlistenport-host-backlog-callback:~:text=If%20host%20is%20omitted%2C%20the%20server%20will%20accept%20connections%20on%20the%20unspecified%20IPv6%20address%20(%3A%3A)%20when%20IPv6%20is%20available%2C%20or%20the%20unspecified%20IPv4%20address%20(0.0.0.0)%20otherwise.