Disable BrowserSync UI by default
We rarely use it.

Disabling it saves a few resources and takes us down to only one URL listed in the BrowserSync ‘Access URLs’ in the console, reducing confusion for users.

Users can still enable it by setting the `ENABLE_BROWSERSYNC_UI` environment variable, and anyone that wants to use it on a regular basis can set the environment variable in their profile.