Fix hover state on focused small radios
The comma is required to separate the two box shadows, but doesn’t have any effect as it is part of the comment. It like this was a regression introduced in a tidy-up as part of 9bc8c33.

Move the comma to the right place to fix the hover state.