Use `aria-hidden` to mask Service Navigation toggle from VoiceOver
`aria-expanded` and `aria-controls` make VoiceOver keep the button
visible in its rotor, so we add `aria-hidden` as a complement to `hidden`.

This is prefered to manually adding and removing each of `aria-expanded` and `aria-controls`
as we'd need to remember their value when resetting them. It also future proofs in case other `aria`
attributes make VoiceOver keep the button in the rotor.