Simplify regex to make it non-polynomial
Thanks to @chrisns who explained:

> “You do not need both `.` and `\S+`. The ambiguity comes from their combination, which makes overlapping matching possible.”

This change does alter the behaviour of the function.

The use of `.\S+` in the current regex allows for a single leading space within the placeholder name (e.g. `%{ foo}`) because:

- `.` matches ANY character except newline, *including* a space
- `\S+` matches one or more *non-whitespace* characters (letters, digits, symbols, etc.)

By removing the `.` in the regular expression a leading space is no longer permitted as part of the placeholder name.

I don’t believe this was an intentional feature, and given the i18n library remains private and I think we can safely make this change as we know how it is used internally.