Prefer Regexp.prototype.test for boolean checks
We don’t need the match that is returned by `String.prototype.match` – we just need a boolean that tells us whether `translation` matches the regex or not.

Use `RegExp.prototype.test` which is the recommended method for this use case. [1]

This also means we can get rid of an eslint-disable comment 🎉

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/match#:~:text=If%20you%20need%20to%20know%20if%20a%20string%20matches%20a%20regular%20expression%20RegExp%2C%20use%20RegExp.prototype.test().