Handle new `unknown` type emited on `pageerror`
Puppeteer recently updated its `pageerror` event types
to no longer be just an `Error`, but also `unknown`. [1^]

This is because JavaScript code may `throw` not just `Error`
objects, but any kind of values. While Puppeteer handles some of the translation,
in some cases it can only forward the value from the browser,
which we then need to handle ourselves. [^2]

For now, we'll just output the value. If we end up seeing `[Array]` or similar
output, we can look at improving the output, but most errors we receive should be `Error` anyways.

[^1]: https://github.com/puppeteer/puppeteer/pull/14349
[^2]: https://github.com/puppeteer/puppeteer/blob/a2eb479ec506947e5912532dc414b37ec2bd7f5e/packages/puppeteer-core/src/cdp/utils.ts#L28