Replace output of `govuk-applied-colour` with `var()` call
Outputs not only the reference to the custom property name,
but also a default value so that components using `govuk-applied-colour`
don't have to worry about importing `core/govuk-frontend-properties`.

We may want to revisit this once we use Sass modules, as we'll be able
to have the applied colour module both:
- declare the list of applied colours
- offer the `govuk-applied-colour`
- output the colours as custom properties

This may ensure that whenever you'd be using `govuk-applied-colour`,
the colours are output.