Remove change of link colour on hover for text and inverse variants
The very subtle change of colour was there to work around a Safari
issue where the change of thickness on hover wouldn't be enough to
trigger a redraw of the link. [1^]

Since then, we've also added a change of `text-decoration-skip-ink`
when the thinckness changes. [2^]

The changes of this property values are monitored by Webkit to trigger
a redraw. [3^].

This means the change of colour is no longer necessary (especially as
the bug is fixed since Safari 16.4 [4^] after a fix in Jan 2023 [5^]).

[1^]: https://github.com/alphagov/govuk-frontend/commit/683591e5980fcc706caf21a9c7fe9bfe46fe85cf
[2^]: https://github.com/alphagov/govuk-frontend/pull/2251
[3^]: https://github.com/WebKit/WebKit/pull/8372/files#diff-dd44b7e08707f7ca39da2c55490b753d5d0b37c3a47a14738748e5fd023fc9daR1240-R1252
[4^]: https://developer.apple.com/documentation/safari-release-notes/safari-16_4-release-notes
[5^]: https://bugs.webkit.org/show_bug.cgi?id=224483