Add helpers to output custom properties based on usage
Introduce a `govuk-property-reference` function and `govuk-export-custom-properties` mixin
working in concert to output custom properties based on their usage.

This allows components imported individually to have the custom properties they use on `:root`.

It works the following way:

```scss
// Wrap CSS that will use custom properties inside the mixin
@include govuk-export-custom-properties {
  .app-component {

    // Use `govuk-property-reference` to get the `var()` call
    // to set as a value, the mixin will take care of outputing a `:root`
    // Usually, this will be encapsulated in a function for accessing design tokens
    color: govuk-property-reference(--_govuk-custom-property, value);
  }
}
```

For now, this does not account for custom properties responsiveness (this will come in a later commit)