Output responsive change to custom properties on demand
This allows for responsive changes to custom properties
to be bundled with the code of the component using them
regardless of if the component is built on its own or
within GOV.UK Frontend as a whole.

It relies on the property value being a map, rather than a plain value.
The actual value is set as the `value` key in that map and a `media`
object allows to list values for various media queries.

This should work with media queries generated with `govuk-from-breakpoint`
and `govuk-until-breakpoint` or any arbitrary media query.