Fix skiplink outline not being visible in forced colours
As skiplinks are usually present right at the top of the viewport and at full width, the outline style we apply to links in forced-colors mode is clipped on three of the four sides. This commit changes the outline-offset in forced-colors so that it is visible within the bounds of the skip link.