Add `aria-hidden` to Service Navigation toggle's initial HTML
`aria-controls` make VoiceOver keep the button visible in its rotor,
on desktop, and keep the button focusable on iPad.

Adding an `aria-hidden` ensures the button is correctly hidden without
introducing a breaking change. This is also the technique we'll use
in the component's JavaScript to hide the button fully as viewport gets resized.